Refactor VulnerabilitySource to be in shared directory access
