+++ exasol/toolbox/tools/security.py
@@ -21,6 +21,7 @@ import typer
 from exasol.toolbox.security import (
     GitHubVulnerabilityIssue,
     VulnerabilityIssue,
+    VulnerabilitySource,
 )
 
 stdout = print
@@ -79,32 +80,6 @@ def from_maven(report: str) -> Iterable[VulnerabilityIssue]:
                 coordinates=dependency_name,
                 references=tuple(references),
             )
-
-
-class VulnerabilitySource(str, Enum):
-    CVE = "CVE"
-    CWE = "CWE"
-    GHSA = "GHSA"
-    PYSEC = "PYSEC"
-
-    @classmethod
-    def from_prefix(cls, name: str) -> VulnerabilitySource | None:
-        for el in cls:
-            if name.upper().startswith(el.value):
-                return el
-        return None
-
-    def get_link(self, package: str, vuln_id: str) -> str:
-        if self == VulnerabilitySource.CWE:
-            cwe_id = vuln_id.upper().replace(f"{VulnerabilitySource.CWE.value}-", "")
-            return f"https://cwe.mitre.org/data/definitions/{cwe_id}.html"
-
-        map_link = {
-            VulnerabilitySource.CVE: "https://nvd.nist.gov/vuln/detail/{vuln_id}",
-            VulnerabilitySource.GHSA: "https://github.com/advisories/{vuln_id}",
-            VulnerabilitySource.PYSEC: "https://github.com/pypa/advisory-database/blob/main/vulns/{package}/{vuln_id}.yaml",
-        }
-        return map_link[self].format(package=package, vuln_id=vuln_id)
 
 
 def identify_pypi_references(
